requirements.txt
google-genai
streamlit
pandas
plotly
python-dotenv
openai
matplotlib
opencv-python-headless
numpy
pillow